requirements.txt
numpy>=1.26.0
pandas>=2.0.0
scipy>=1.11.0
matplotlib>=3.7.0
python-dateutil>=2.8.2
requests>=2.31.0
